Correctly read blank cells in MS Word with NVDA's UIA for MS Word option enabled (#12731)
When using UIA to access Microsoft word documents with NVDA, NVDA failed to announce row and column numbers when arrowing onto blank table cells. Also, navigating a table with the control+alt+arrow keys and landing on blank cells would sometimes announce leaving the table.
This was caused by the fact that when expanding a IUIAutomationTextRange in a blank table cell to any text unit, the range will cover the end-of-cell marker, and therefore span more than the cell itself, thus it looks like the range is not entirely inside the cell.

Description of how this pull request fixes the issue:
Work around this MS Word bug by detecting when in this situation: the range is not collapsed, yet the text in the range is either empty or the end-of-cell marker. And if so, Use a copy of the range, that is collapsed to the start (entirely inside the cell) to fetch the text and fields.
This pr also makes a change to Braille, so that when building a line of rich text in braille, a TextInfo's getTextWithFields method is only called if the textInfo is not collapsed. this stops repeating of fields for collapsed ranges, which was sometimes seen in Microsoft Word with UIA enabled.

Fixes #11043
Fixes #11457
Fixes #12490