Option to configure when the error sound may be played. (PR #12689)
Fixes #12672

# Summary of the issue:
The error sound that is played in test versions may also be useful to testers on stable and RC releases.
The motivation to have this feature included in NVDA's core rather than in an add-on is that it may be useful to test NVDA stable versions with add-ons disabled.

# Description of how this pull request fixes the issue:
Configurable via 'Play a sound for logged errors' in the advanced settings panel.

- 'Only in test versions': NVDA plays error sounds only in NVDA test versions (alpha, beta and from source) but not in RC or stable.
- 'Yes': Error sound for all version (test or stable)

Discussion on #12672, recommends not provide an option to disable error sounds in tests versions, since test versions are aimed to capture potential issues. Thus error notifications should not be disabled for this usage.

If an error occurs before the config is initialized, the default behavior is applied, i.e. the error sound is played only in test versions.

Co-authored-by: Łukasz Golonka <lukasz.golonka@mailbox.org>
Co-authored-by: Leonard de Ruijter <leonardder@users.noreply.github.com>
Co-authored-by: Reef Turner <reef@nvaccess.org>