Split script for reading status bar into  two - one for reading and one for moving review cursor into its location (#12771)
Link to issue number:
Closes #8600
Implements part of #7314

Summary of the issue:
Currently when reading status bar with NVDA+END review cursor is unconditionally moved to its location. This is inconsistent with all other commands for retrieving information and aside from that is extremely annoying for people working with review cursor now following the caret (I've been in a situation in which I've left review cursor somewhere either because I wanted to monitor a given object for changes, or because I was following some documentation with review cursor in a separate window and wanted to avoid unnecessary alt tabbing and lost my review position because I had to check something on a status bar countless times).

Description of how this pull request fixes the issue:
Review cursor is no longer moved to the location of the status bar. There is a separate script (no gesture assigned by default) which moves review cursor there. As described in #7314 I took this opportunity to split script for reading status bar into a separate ones for reading, spelling and copying.

Testing strategy:
With NVDA+end made sure that first press still reads, second spells and third copies. Made sure that after each of these operations review cursor has not been moved. Tested each of the newly introduced scripts made sure they work. Compiled user guide - ensured it renders as expected.