Descend into children for dialog text for option pane
As is done for objects of role 'controlTypes.Role.PANE'
also descend into children when retrieving the dialog text
from an object of role 'controlTypes.Role.OPTIONPANE'.

This is used e.g. in LibreOffice's message dialogs and
makes sure that the text in the dialogs is read out.

Note: In addition to this change in NVDA, it also requires
the following commit recently committed to LibreOffice [1]
(which will be contained in LibreOffice 7.3)

    commit e29f80efc826fdc8a8e4b7cbdd3becf2f074cce4
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Sep 1 16:25:01 2021 +0200

        tdf#135588 wina11y: Map AccessibleRole::STATIC to ROLE_SYSTEM_STATICTEXT

Its commit message has some more background information.

Fixes #11687

[1] https://git.libreoffice.org/core/+/e29f80efc826fdc8a8e4b7cbdd3becf2f074cce4%5E%21