Restart nvda_dmp on exception (#12827)
In some situations, nvda_dmp stops functioning. NVDA falls back to Difflib in these cases, so the console continues to function, but frequent error sounds are played and Difflib is forceably used until NVDA is restarted.

Description of how this pull request fixes the issue:
When nvda_dmp fails to diff, restart it.