Descend into children for dialog text for option pane (Fixes #11687) (#12801)
Text in LibreOffice message dialogs is not announced by NVDA.
Description of how this pull request fixes the issue:
This commit makes sure that the children of objects of role 'controlTypes.Role.OPTIONPANE' are taken into account when determining the text to announce for a dialog, as is done for objects of role 'controlTypes.Role.PANE'.
For LibreOffice (message) dialogs, the dialog text are objects of role 'controlTypes.Role.STATICTEXT' that are children of an object of type 'controlTypes.Role.OPTIONPANE'.
Those were previously not considered when determining the dialog text to announce. This change ensures that they are now and that the actual dialog text is thus announced by NVDA, just as is done (and has been the case) for objects of role 'controlTypes.Role.PANE'.