Use UIA to access MS Word documents by default for MS Office build 16.0.13901 and above (#12770)
Microsoft Word 2016 exposes a rich UI Automation implementation. For some time now, users have been able to optionally turn this on with an advanced setting. NVDA's support for MS Word via UIA has major performance advantages over the older object model support, so NVDA should use the UIA support by default where available. However, as the UIA implementation improved throughout Office 2016's lifetime, we should only enable our support for recent builds of Office 2016, specifically for build 13901 and higher.

Description of how this pull request fixes the issue:
For a _WwG window that is found to have a native UI Automation implementation, the only situation where this UIA implementation will be ignored and therefore NVDA will fall back to the Office object model is: If it is an MS Office app, the build is < 13901, NVDA is able to inject in-process, and the user has not turned on Always use UI Automation to access Microsoft Word document controls.
The user guide has been updated to note that UIA will be used by default for MS Word version 13901 and higher.