Ensure that NVDA no longer checks validity of input for non contracted braille tables when in text controls (#12850)
Fixes #12667

Summary of the issue:
When entering text from a braille display NVDA reports that it is unsupported even though entered combination is valid. This check has been introduced in #7843 to prevent a situation in which someone uses contracted table outside text content and enters a combination which cannot be converted to a single character (a good example is entering dots 356 using English UK Grade 2 which is a contraction for "was"). The check however was too broad and didn't take into account that for literary / computer braille tables single combination of dots can either always be converted to a single character, or if not it does not result in multiple characters i.e. number sign.

Description of how this pull request fixes the issue:
Now the check applies only when the table in use is contracted.