Scrolling braille in MS Word with a Chinese input method enabled no longer incorrectly jumps back (#12868)
Fixes #12855

Summary of the issue:
When in Microsoft Word, and using a Chinese input method such as Taiwan - Microsoft Quick, and NVDA is accessing Microsoft Word via UI Automation, trying to scroll forward through the document with a braille display causes the braille display to keep jumping back to the original caret position.
This is due to the fact that while the Chinese input method is enabled, legacy MSAA caret events are fired on the Word document window. NVDA is not ignoring these, even though NVDA is configured to access Microsoft word with UIA.

Description of how this pull request fixes the issue:
In IAccessibleHandler.processGenericWinEvent: if an MSAA caret event is identified, before we route this onto the focused object, we first check if the focused object is in deed IAccessible (MSAA), and if it is not (E.g. it is UIA) then we drop the event completely.
This pr also improves MSAA debug logging a little by including the winEvent info in the message about the caret being routed to the focus, and also routes generic events for the focus to the existing focus if their event object is not already the existing focus. This really only stops some extra redundant logging.