MS Word with UIA: support moving by sentence with legacy object model when available (#12874)
Fixes #9254

Summary of the issue:
In the past, NVDA has had the functionality to move by sentence in MS word (alt+downArrow and alt+upArrow). This functionality used the MS Word object model.
With the switch to using UI automation to access MS Word, move by sentence was lost, as MS Word's UI automation implementation provides no concept of a sentence unit.

Description of how this pull request fixes the issue:
This pr provides an implementation of moving by sentence for MS Word with UIA by falling back to the MS word object model for this specific feature, if the object model is available.
This therefore will work in MS Word and Outlook, but not in Windows Mail.