Better detection of spelling errors when typing in MS Word with UIA (#12925)
Fixes #12161

Summary of the issue:
When in any editable text control (E.g. a Word document), each time the user completes a word by typing space or another punctuation character, NVDA checks the formatting of the completed word to see if it is marked as a spelling error, and if so plays a special text error buzz sound.
However, in at least MS Word with UIA, this check for spelling errors never really finds an error as MS Word has not yet actually detected the error itself.
We need to slow down NVDA's detection to give the app more time.

Description of how this pull request fixes the issue:
Delay the fetch and check of the completed word's formatting into a future core cycle (specifically a 50 ms delay) to give the app more time to detect the error itself.