Support table navigation for WPF list item DataItem objects (#12928)
Steps to reproduce the issue
Open PowerShell
Type Get-Process | Out-GridView
Observe that table navigation doesn't work on the several entries in the grid.

Description of how this pull request fixes the issue:
This pr adds the RowWithFakeNavigation on these objects, thereby automatically exposing the ability to table nav through the grid, including reading of column headers.

Testing strategy:
Tested in both PowerShell 5.1 and PowerShell Core 7.1

Known issues with pull request:
This change might hit untested controls. Note that the logic in RowWithFakeNavigation continuously checks child count, so I think it's pretty safe to add. It is also added to icons on the desktop, for example.