Fix reading order / description spoken during quick nav (PR #12933)
Fixes #12751
Broken in #12500

When using quick nav ("h" in browse mode) to navigate to a heading nested in an article with a description, the description was announced before the heading. For quicknav and for focus changes, the most inner element should always be announced first, working outwards to provide additional context.