Ensure that clipboard manager does not steal focus when starting  MS Office programs (#12736)
As described in #4199 when starting Office programs Word, Excel clipboard manager can sometimes steal accessibility focus making it impossible to interact with the document / spreadsheet.
Most cases were fixed but it seems one has been missed - focus lands on something called Collect and Paste 2.0 ,making interaction with the actual document difficult.

Description of how this pull request fixes the issue:
Since the "Collect and paste" has a custom eventObjectID SDM~has not been applied to it causing focus to be incorrect. This custom ID has been added to the list of objects for which SDM~applies.