Improve selection and merged cell announcements in LibreOffice Calc 7.3 and above (#12849)
Link to issue number:
Fixes #9310
Fixes #6897

Summary of the issue:
In LibreOffice calc, selection is not announced and for merged cells, only the first cell is announced.

Description of how this pull request fixes the issue:
In LibreOffice 7.3 development branch, support for IAccessibleTable2 and IAccessibleTableCell were added. This allows us to implement something that works as expected.