Honor Windows setting to swap primary mouse button (#12922)
Fixes #12642

Summary of the issue:
NVDA may at times resort to clicking the mouse to perform particular actions on behalf of the user, such as activating a link or bringing up a context menu.
Currently, NVDA clicks the left mouse button to activate / select items, and clicks the right mouse button to display a context menu. This is generally expected behaviour.
However, Windows has a user setting allowing the user to swap the mouse buttons, such that the primary button (activating / selecting) is the right mouse button, and the secondary (context menu) is the left.
If the user has chosen to swap the buttons, NVDA may inadvertantly display a context menu when trying to activate an item, or the other way round.

Description of how this pull request fixes the issue:
This pr adds doPrimaryClick and doSecondaryClick functions to mouseHandler that honor the Windows user preference for which button is the primary mouse button.
All code in NVDA that licks aa mouse button in order to perform activation, selection or show a context menu has been changed to call these new functions in place of calling executeMouseEvent for the left or right mouse button.
However, The mouse click scripts in globalCommands have not been changed, as these are specifically named after which physical mouse button they click, rather than performing a logical primary or secondary action.