Ensure that NVDA can get file version information for binaries in system32 on 64-bit versions of Windows (#12943)
discussion in #12852

Summary of the issue:
NVDA can retrieve information for a binary file such as application name and version from which given appModule has been created. While they're not shown to the user directly they can be inspected as part of the developer info - also NVDA relies on them for various stuff internally.

Description of how this pull request fixes the issue:
For a duration of retrieving file info for a binary placed in system32 and when running on a version of Windows which have both system32 and SysWOW64 WOW64 redirection is disabled. While at it I've removed deprecated shlobj.SHGetFolderPath and replaced it with SHGetKnownFolderPath and removed unused function from the config module ((while not strictly related it relied on SHGetFolderPath).

Testing strategy:
On 64-bit version of Windows installed self signed build from this branch, enabled service debug, started an application with the admin privileges, inspected developer info for the UAC screen and made sure that for consent.exe file version and file name can be retrieved. Made sure that this code is not executed on 32-bit version of Windows and that file version info can still be retrieved there.

Co-authored-by: buddsean <sean@nvaccess.org>