Fixup FolderId usage and change log (#12986)
Fixup of #12943

Summary of the issue:
Changelog entries were missing for #12943

UpperCamelCase is the standard for class names, including enums, however FOLDERID was used as the casing for the enum, which mirrors the windows constants.
There is a similar issue for the enum members (which used UpperCamelCase instead of CAP_SNAKE_CASE).

If we want to change our practice to be consistent with what an Enum represents, we should update codingStandards.md

Also the enum type was not being fully leveraged by SHGetKnownFolderPath, requiring .value to be used unnecessarily when calling the function.

Description of how this pull request fixes the issue:
Updates the casing of the enum and it's members. Updates the change log, and fixes up some earlier entries.