Use `casefold` when comparing file paths as `os.path.commonPath` fails to handle paths from different disks (#13009)
PR #12943 improved code which retrieves version information for a given executable files placed in system32 under 64-bit versions of Windows. To do so it is necessary to check if path of the given binary starts with path of system32 directory for the current system. The recommended way to compare path's in a case insensitive way was os.pathcommonPath`, however it turns out it does not work when path's are from different drives.
In case of reporter of #13008 they had Office installed in a custom location probably on non system disk.

Description of how this pull request fixes the issue:
This PR simplifies the code to use standard str.casefold for comparisons.